backend/pyproject.toml
[tool.poetry]
name = "cro-agent-backend"
version = "0.1.0"
description = "Backend API for Shopify CRO agent"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100"
uvicorn = {extras = ["standard"], version = "^0.23"}
python-dotenv = "^1.1"
celery = "^5.3"
redis = "^4.5"
shopifyapi = "^10.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"